Add files via upload
These are all the projects I made with java.
Most of them were created to give me extra practice with the language. Many of these .java files are "runners" to run other .java files (both of which should be included). The majority of these programs work, though not all, and they vary in complexity.